Add styled MUI and extract styling into separate files
