+++ src/components/index.ts
@@ -2,7 +2,13 @@ import { GenericErrorMessage } from "./GenericErrorMessage";
 import { OccupationDropdown } from "./FormElements/OccupationDropdown";
 import { StateDropdown } from "./FormElements/StateDropdown";
 import { Loading } from "./Loading";
-import { LastNameField } from "./FormElements";
+import { Form } from "./Form";
+import {
+  LastNameField,
+  FirstNameField,
+  EmailField,
+  PasswordField,
+} from "./FormElements";
 
 export {
   GenericErrorMessage,
@@ -10,4 +16,8 @@ export {
   StateDropdown,
   Loading,
   LastNameField,
+  FirstNameField,
+  EmailField,
+  PasswordField,
+  Form,
 };
